LayoutTests/imported/w3c/web-platform-tests/dom/nodes/Document-createEvent-touchevent.window-expected.txt


PASS document.createEvent('TouchEvent') should throw if 'expose legacy touch event APIs' is false
PASS document.createEvent('touchevent') should throw if 'expose legacy touch event APIs' is false
PASS document.createEvent('TOUCHEVENT') should throw if 'expose legacy touch event APIs' is false
